+++ public/index.html
@@ -46,7 +46,7 @@
               value="beginner"
               checked
               />
-              <label for="beginner">Beginner</label>
+              <label for="beginner">1</label>
             </div>
           <div class="radio">
               <input
@@ -55,7 +55,7 @@
               name="difficulty"
               value="intermediate"
               />
-            <label for="intermediate">Intermediate</label>
+            <label for="intermediate">2</label>
           </div>
           <div class="radio">
             <input
@@ -64,7 +64,7 @@
               name="difficulty"
               value="advanced"
             />
-            <label for="advanced">Advanced</label>
+            <label for="advanced">3</label>
           </div>
         </div>
       </div>
